split up toolbar
